Fixes for ophyd async types mega merge (#593)
* Fix AperturePosition radius for robot load to be 0 instead of None
* Make MockBlock call superclass init()
* Fix usage of DatasetTable
* Remove timeout parameter from get_mock_put() asserts
* Fix Thawing tests
* Fix zebra mock using now-removed _backend private implementation
* Fix thawing plan borkages due to source enum changes and default parameter injection fallout
* Change flyscan xrc to use the new-style XrcResult
* Fix pin tip centring to be in line with new types
* fix zebra arming mock problem
* Fix race condition in grid detect unit tests
* Fix typing issue with pin-tip-centring tests
* Fix OAV changes to work with ophyd-async typing changes
* Fix test_chip_manager.py for ophyd-async type changes
* Add test coverage for pin-tip-detection return on exhaustion

+++ tests/system_tests/conftest.py
@@ -139,7 +139,7 @@ def oav_for_system_test(test_config_files):
     set_mock_value(oav.grid_snapshot.y_size, 768)
     set_mock_value(oav.grid_snapshot.top_left_x, 50)
     set_mock_value(oav.grid_snapshot.top_left_y, 100)
-    size_in_pixels = int(0.1 * 1000 / 1.25)
+    size_in_pixels = 0.1 * 1000 / 1.25
     set_mock_value(oav.grid_snapshot.box_width, size_in_pixels)
     unpatched_snapshot_trigger = oav.grid_snapshot.trigger
